Rework documentation
