+++ ARM.CMSIS-Compiler.pdsc
@@ -126,7 +126,7 @@
         #define RTE_Compiler_File_Interface_Custom /* Compiler File Interface: Custom */
       </RTE_Components_h>
       <files>
-        <file category="source"  name="template/retarget_fs.c" attr="template" select="File Interface Custom Template"/>
+        <file category="source"  name="template/file_interface/retarget_fs.c" attr="template" select="File Interface Custom Template"/>
       </files>
     </component>
 
@@ -186,9 +186,8 @@
         <file category="source" name="source/armcc/retarget_io.c"     condition="ARMCC"/>
         <file category="source" name="source/gcc/retarget_syscalls.c" condition="GCC"/>
 
-        <file category="source" name="template/stdin_user.c"     attr="template" select="STDIN User Template"/>
-        <file category="source" name="template/stdin_USART.c"    attr="template" select="STDIN via USART"/>
-        <file category="source" name="template/stdin_keyboard.c" attr="template" select="STDIN via Keyboard"/>
+        <file category="source" name="template/io/stdin_user.c"     attr="template" select="STDIN User Template"/>
+        <file category="source" name="template/io/stdin_USART.c"    attr="template" select="STDIN via USART"/>
       </files>
     </component>
     <component Cclass="CMSIS-Compiler" Cgroup="I/O" Csub="STDOUT" Cvariant="Breakpoint" Cversion="1.0.0" condition="CortexDevice">
@@ -235,9 +234,9 @@
         <file category="source" name="source/armcc/retarget_io.c"     condition="ARMCC"/>
         <file category="source" name="source/gcc/retarget_syscalls.c" condition="GCC"/>
 
-        <file category="source" name="template/stdout_user.c"    attr="template" select="STDOUT User Template"/>
-        <file category="source" name="template/stdout_USART.c"   attr="template" select="STDOUT via USART"/>
-        <file category="source" name="template/stdout_display.c" attr="template" select="STDOUT via Display"/>
+        <file category="source" name="template/io/stdout_user.c"    attr="template" select="STDOUT User Template"/>
+        <file category="source" name="template/io/stdout_USART.c"   attr="template" select="STDOUT via USART"/>
+        <file category="source" name="template/io/stdout_display.c" attr="template" select="STDOUT via Display"/>
       </files>
     </component>
     <component Cclass="CMSIS-Compiler" Cgroup="I/O" Csub="STDERR" Cvariant="Breakpoint" Cversion="1.0.0" condition="CortexDevice">
@@ -273,9 +272,9 @@
         <file category="source" name="source/armcc/retarget_io.c"     condition="ARMCC"/>
         <file category="source" name="source/gcc/retarget_syscalls.c" condition="GCC"/>
 
-        <file category="source" name="template/stderr_user.c"    attr="template" select="STDERR User Template"/>
-        <file category="source" name="template/stderr_USART.c"   attr="template" select="STDERR via USART"/>
-        <file category="source" name="template/stderr_display.c" attr="template" select="STDERR via Display"/>
+        <file category="source" name="template/io/stderr_user.c"    attr="template" select="STDERR User Template"/>
+        <file category="source" name="template/io/stderr_USART.c"   attr="template" select="STDERR via USART"/>
+        <file category="source" name="template/io/stderr_display.c" attr="template" select="STDERR via Display"/>
       </files>
     </component>
 
@@ -309,7 +308,7 @@
         <file category="header" name="include/retarget_tty.h"/>
         <file category="source" name="source/armcc/retarget_io.c"/>
 
-        <file category="source" name="template/tty_user.c" attr="template" select="TTY User Template"/>
+        <file category="source" name="template/io/tty_user.c" attr="template" select="TTY User Template"/>
       </files>
     </component>
 
@@ -321,9 +320,9 @@
         #define RTE_Compiler_OS_Interface_Custom /* Compiler OS Interface: Custom */
       </RTE_Components_h>
       <files>
-        <file category="source" name="template/armcc/retarget_os.c"     condition="ARMCC" attr="template" select="OS Interface Custom Template"/>
-        <file category="source" name="template/gcc/retarget_lock.c"     condition="GCC"   attr="template" select="OS Interface Lock Custom Template"/>
-        <file category="source" name="template/gcc/retarget_syscalls.c" condition="GCC"   attr="template" select="OS Interface SysCalls Custom Template"/>
+        <file category="source" name="template/os_interface/armcc/retarget_os.c"     condition="ARMCC" attr="template" select="OS Interface Custom Template"/>
+        <file category="source" name="template/os_interface/gcc/retarget_lock.c"     condition="GCC"   attr="template" select="OS Interface Lock Custom Template"/>
+        <file category="source" name="template/os_interface/gcc/retarget_syscalls.c" condition="GCC"   attr="template" select="OS Interface SysCalls Custom Template"/>
       </files>
     </component>
     <component Cclass="CMSIS-Compiler" Cgroup="OS Interface" Csub="CMSIS-RTOS2" Capiversion="1.0.0" Cversion="1.0.0" condition="CortexDevice CMSIS-RTOS2">
