+++ doxygen/RetargetIO.dxy.in
@@ -44,7 +44,7 @@ PROJECT_NUMBER         = "Version {projectNumber}"
 # for a project that appears at the top of each page and should give viewer a
 # quick idea about the purpose of the project. Keep the description short.
 
-PROJECT_BRIEF          = "Debugger Views for Status and Event Information"
+PROJECT_BRIEF          = "Standard C Library File, I/O and OS Retargeting"
 
 # With the PROJECT_LOGO tag one can specify a logo or an icon that is included
 # in the documentation. The maximum height of the logo should not exceed 55
@@ -187,7 +187,7 @@ SHORT_NAMES            = NO
 # description.)
 # The default value is: NO.
 
-JAVADOC_AUTOBRIEF      = NO
+JAVADOC_AUTOBRIEF      = YES
 
 # If the JAVADOC_BANNER tag is set to YES then doxygen will interpret a line
 # such as
@@ -871,11 +871,20 @@ WARN_LOGFILE           =
 # spaces. See also FILE_PATTERNS and EXTENSION_MAPPING
 # Note: If this tag is empty the current directory is searched.
 
-INPUT                  = src/compiler.txt \
-                         ../example/README.md \
-                         src/files.txt \
-                         src/history.txt \
-                         ../source/armcc/retarget_io.c
+INPUT                  = src/main.txt                   \
+                         src/history.txt                \
+                         src/overview.txt               \
+                         src/rt_io.txt                  \
+                         src/rt_os.txt                  \
+                         ../example/README.md           \
+                         src/templates.txt              \
+                         src/retarget_os.txt            \
+                         src/retarget_os_armclib.txt    \
+                         src/retarget_os_newlib.txt     \
+                         src/retarget_fs.txt            \
+                         ../include/armcc/retarget_os.h \
+                         ../include/gcc/retarget_os.h   \
+                         ../include/retarget_fs.h
 
 # This tag can be used to specify the character encoding of the source files
 # that doxygen parses. Internally doxygen uses the UTF-8 encoding. Doxygen uses
@@ -964,7 +973,7 @@ RECURSIVE              = NO
 # Note that relative paths are relative to the directory from which doxygen is
 # run.
 
-EXCLUDE                =
+EXCLUDE                = ../template/*
 
 # The EXCLUDE_SYMLINKS tag can be used to select whether or not files or
 # directories that are symbolic links (a Unix file system feature) are excluded
@@ -997,9 +1006,9 @@ EXCLUDE_SYMBOLS        =
 # that contain example code fragments that are included (see the \include
 # command).
 
-EXAMPLE_PATH           = ../source \
-                         ../template \
-                         ../example
+EXAMPLE_PATH           = ../example  \
+                         ../source   \
+                         ../template
 
 # If the value of the EXAMPLE_PATH tag contains directories, you can use the
 # EXAMPLE_PATTERNS tag to specify one or more wildcard pattern (like *.cpp and
