--- doxygen/src/examples.txt
@@ -1,20 +1,0 @@
-/* ========================================================================================================================== */
-/**
-\page Retarget_Examples Examples
-
-This software pack contains an example project that shows how to do I/O retargeting.
-
-The following  example project is available:
-
-Example Project                  | Description
-:--------------------------------|:-----------------------------------------
-\subpage Retarget_Examples_UART  | Shows how to use a U(S)ART for STDOUT.
-*/
-
-/* ========================================================================================================================== */
-/**
-\page Retarget_Examples_UART  Retarget Input and Output via UART
-
-\include ../../example/README.md
-
-*/
