+++ doxygen/src/main.txt
@@ -1,0 +1,25 @@
+/**
+\mainpage
+\b CMSIS-Compiler provides software components that simplify the standard C run-time library retargeting.\n
+
+
+Using CMSIS-Compiler developers can retarget standard I/O streams and file operations to specific platform and enable
+thread safe operations using RTOS interface.
+
+\section doc_content Content
+This documentation contains the following sections:
+
+- \subpage rev_hist : lists CMSIS-Compiler releases.
+- \ref rt_overview : provides an overview about the CMSIS-Compiler software component.
+- \ref rt_io : explains low-level I/O retargeting, list available components and describes available user code templates.
+- \ref rt_os : explains multithreading support retargeting, list available components and describes available user code templates.
+- The [I/O Retarget example project](../example/README.md) shows how to retarget the output to a UART on an Arm Virtual Hardware model.
+- \ref rt_templates : contains the user code template files for the different use cases.
+
+\note
+
+The \b CMSIS-Compiler component is available for
+<a href="https://developer.arm.com/Tools%20and%20Software/Arm%20Compiler%20for%20Embedded" target="_blank">Arm Compiler for Embedded</a>
+and the <a href="https://developer.arm.com/Tools%20and%20Software/GNU%20Toolchain" target="_blank">Arm GNU Toolchain (GCC)</a>.
+IAR Compiler support will be added in a future revision of the pack. 
+*/
