+++ template/io/stderr_USART.c
@@ -1,6 +1,6 @@
 /*-----------------------------------------------------------------------------
- * Name:    stdout_USART.c
- * Purpose: STDOUT USART Template
+ * Name:    stderr_USART.c
+ * Purpose: STDERR USART Template
  * Rev.:    1.0.0
  *-----------------------------------------------------------------------------*/
  
@@ -22,19 +22,19 @@
  * limitations under the License.
  */
  
-#include "retarget_stdout.h"
+#include "retarget_stderr.h"
 #include "Driver_USART.h"
  
 //-------- <<< Use Configuration Wizard in Context Menu >>> --------------------
  
-// <h>STDOUT USART Interface
+// <h>STDERR USART Interface
  
 //   <o>Connect to hardware via Driver_USART# <0-255>
 //   <i>Select driver control block for USART interface
 #define USART_DRV_NUM           0
  
 //   <o>Baudrate
-#define USART_BAUDRATE          9600
+#define USART_BAUDRATE          115200
  
 // </h>
  
@@ -47,11 +47,11 @@ extern ARM_DRIVER_USART  USART_Driver_(USART_DRV_NUM);
  
  
 /**
-  Initialize stdout
+  Initialize stderr
  
   \return          0 on success, or -1 on error.
 */
-int stdout_init (void) {
+int stderr_init (void) {
   int32_t status;
  
   status = ptrUSART->Initialize(NULL);
@@ -67,21 +67,21 @@ int stdout_init (void) {
                              ARM_USART_FLOW_CONTROL_NONE,
                              USART_BAUDRATE);
   if (status != ARM_DRIVER_OK) return (-1);
-
+ 
   status = ptrUSART->Control(ARM_USART_CONTROL_TX, 1);
   if (status != ARM_DRIVER_OK) return (-1);
-
+ 
   return (0);
 }
  
  
 /**
-  Put a character to the stdout
+  Put a character to the stderr
  
   \param[in]   ch  Character to output
   \return          The character written, or -1 on write error.
 */
-int stdout_putchar (int ch) {
+int stderr_putchar (int ch) {
   uint8_t buf[1];
  
   buf[0] = ch;
