+++ template/io/stdin_user.c
