+++ template/io/stdout_user.c
