+++ template/os_interface/gcc/retarget_lock.c
@@ -1,5 +1,5 @@
 /*-----------------------------------------------------------------------------
- * Name:    retarget_lock_user.c
+ * Name:    retarget_lock.c
  * Purpose: Locking Routines Retarget Template
  * Rev.:    1.0.0
  *-----------------------------------------------------------------------------*/
