--- template/stderr_display.c
@@ -1,56 +1,0 @@
-/*-----------------------------------------------------------------------------
- * Name:    stderr_display.c
- * Purpose: STDERR Display Template
- * Rev.:    1.0.0
- *-----------------------------------------------------------------------------*/
- 
-/*
- * Copyright (C) 2023 ARM Limited or its affiliates. All rights reserved.
- *
- * SPDX-License-Identifier: Apache-2.0
- *
- * Licensed under the Apache License, Version 2.0 (the License); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an AS IS BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
- 
-#include "retarget_stderr.h"
-#include "GUI.h"
- 
- 
-/**
-  Initialize stderr
- 
-  \return          0 on success, or -1 on error.
-*/
-int stderr_init (void) {
- 
-  GUI_Init();
- 
-  GUI_SetBkColor(GUI_BLACK);
-  GUI_Clear();
- 
-  GUI_SetColor(GUI_WHITE);
- 
-  return (0);
-}
- 
- 
-/**
-  Put a character to the stderr
- 
-  \param[in]   ch  Character to output
-  \return          The character written, or -1 on write error.
-*/
-int stderr_putchar (int ch) {
-  GUI_DispChar((unsigned short)ch);
-  return (ch);
-}
